+++ core/src/main/java/com/github/starnowski/jamolingo/core/mapping/NavigationMapping.java
@@ -8,6 +8,10 @@ import java.util.Objects;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class NavigationMapping {
 
+  /** Target MongoDB database name */
+  @JsonProperty("database")
+  private String database;
+
   /** Target MongoDB collection name */
   @JsonProperty("collection")
   private String collection;
@@ -19,6 +23,35 @@ public class NavigationMapping {
   /** Field in the target collection used for join */
   @JsonProperty("foreignField")
   private String foreignField;
+
+  /**
+   * Returns target MongoDB database name.
+   *
+   * @return database name
+   */
+  public String getDatabase() {
+    return database;
+  }
+
+  /**
+   * Sets target MongoDB database name.
+   *
+   * @param database database name
+   */
+  public void setDatabase(String database) {
+    this.database = database;
+  }
+
+  /**
+   * Sets target MongoDB database name.
+   *
+   * @param database database name
+   * @return current instance
+   */
+  public NavigationMapping withDatabase(String database) {
+    this.database = database;
+    return this;
+  }
 
   /**
    * Returns target MongoDB collection name.
@@ -112,20 +145,24 @@ public class NavigationMapping {
     if (this == o) return true;
     if (o == null || getClass() != o.getClass()) return false;
     NavigationMapping that = (NavigationMapping) o;
-    return Objects.equals(collection, that.collection)
+    return Objects.equals(database, that.database)
+        && Objects.equals(collection, that.collection)
         && Objects.equals(localField, that.localField)
         && Objects.equals(foreignField, that.foreignField);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(collection, localField, foreignField);
+    return Objects.hash(database, collection, localField, foreignField);
   }
 
   @Override
   public String toString() {
     return "NavigationMapping{"
-        + "collection='"
+        + "database='"
+        + database
+        + '\''
+        + ", collection='"
         + collection
         + '\''
         + ", localField='"
@@ -148,9 +185,21 @@ public class NavigationMapping {
 
   /** Builder for NavigationMapping. */
   public static class Builder {
+    private String database;
     private String collection;
     private String localField;
     private String foreignField;
+
+    /**
+     * Sets target MongoDB database name.
+     *
+     * @param database database name
+     * @return builder
+     */
+    public Builder withDatabase(String database) {
+      this.database = database;
+      return this;
+    }
 
     /**
      * Sets target MongoDB collection name.
@@ -192,6 +241,7 @@ public class NavigationMapping {
      */
     public NavigationMapping build() {
       NavigationMapping navigationMapping = new NavigationMapping();
+      navigationMapping.setDatabase(database);
       navigationMapping.setCollection(collection);
       navigationMapping.setLocalField(localField);
       navigationMapping.setForeignField(foreignField);