+++ core/src/main/java/com/github/starnowski/jamolingo/core/mapping/ODataMongoMappingFactory.java
@@ -53,6 +53,7 @@ public class ODataMongoMappingFactory {
 
     EntityMapping entity = new EntityMapping();
 
+    entity.setDatabase(entityType.getNamespace());
     entity.setCollection(defaultCollectionName(entityType));
     entity.setProperties(mapProperties(entityType));
     return entity;
@@ -92,6 +93,7 @@ public class ODataMongoMappingFactory {
     pm.setType(navProp.getType().getNamespace() + "." + navProp.getType().getName());
 
     NavigationMapping nm = new NavigationMapping();
+    nm.setDatabase(navProp.getType().getNamespace());
     nm.setCollection(navProp.getType().getName()); // Default to EntityType name
     // nm.setLocalField(...); // Can be inferred or left for manual override
     // nm.setForeignField("_id"); // Default convention